removes base executor
