+++ lib/Benchmark/Executor/Method/template/execute_static_methods.template
