--- lib/Benchmark/Executor/MicrotimeExecutor.php
@@ -1,90 +1,0 @@
-<?php
-
-/*
- * This file is part of the PHPBench package
- *
- * (c) Daniel Leech <daniel@dantleech.com>
- *
- * For the full copyright and license information, please view the LICENSE
- * file that was distributed with this source code.
- *
- */
-
-namespace PhpBench\Benchmark\Executor;
-
-use PhpBench\Benchmark\Metadata\SubjectMetadata;
-use PhpBench\Benchmark\Remote\Launcher;
-use PhpBench\Benchmark\Remote\Payload;
-use PhpBench\Model\Iteration;
-use PhpBench\Model\Result\MemoryResult;
-use PhpBench\Model\Result\TimeResult;
-use PhpBench\Registry\Config;
-use Symfony\Component\OptionsResolver\OptionsResolver;
-
-/**
- * This class generates a benchmarking script and places it in the systems
- * temp. directory and then executes it. The generated script then returns the
- * time taken to execute the benchmark and the memory consumed.
- */
-class MicrotimeExecutor implements BenchmarkExecutorInterface
-{
-    /**
-     * @var Launcher
-     */
-    private $launcher;
-
-    public function __construct(Launcher $launcher)
-    {
-        $this->launcher = $launcher;
-    }
-
-    public function execute(SubjectMetadata $subjectMetadata, Iteration $iteration, Config $config): void
-    {
-        $tokens = [
-            'class' => $subjectMetadata->getBenchmark()->getClass(),
-            'file' => $subjectMetadata->getBenchmark()->getPath(),
-            'subject' => $subjectMetadata->getName(),
-            'revolutions' => $iteration->getVariant()->getRevolutions(),
-            'beforeMethods' => var_export($subjectMetadata->getBeforeMethods(), true),
-            'afterMethods' => var_export($subjectMetadata->getAfterMethods(), true),
-            'parameters' => var_export($iteration->getVariant()->getParameterSet()->getArrayCopy(), true),
-            'warmup' => $iteration->getVariant()->getWarmup() ?: 0,
-        ];
-        $payload = $this->launcher->payload(__DIR__ . '/template/microtime.template', $tokens);
-
-        $this->launch($payload, $iteration, $config);
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function launch(Payload $payload, Iteration $iteration, Config $options)
-    {
-        $phpConfig = [
-            'max_execution_time' => 0,
-        ];
-
-        $payload->mergePhpConfig($phpConfig);
-        $result = $payload->launch();
-
-        if (isset($result['buffer']) && $result['buffer']) {
-            throw new \RuntimeException(sprintf(
-                'Benchmark made some noise: %s',
-                $result['buffer']
-            ));
-        }
-
-        $iteration->setResult(new TimeResult($result['time']));
-        $iteration->setResult(MemoryResult::fromArray($result['mem']));
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function configure(OptionsResolver $options)
-    {
-        $options->setDefaults([
-            'php_config' => [],
-        ]);
-    }
-}
