+++ lib/Extension/CoreExtension.php
@@ -18,9 +18,9 @@ use PhpBench\Assertion\ComparatorAsserter;
 use PhpBench\Benchmark\BaselineManager;
 use PhpBench\Benchmark\BenchmarkFinder;
 use PhpBench\Benchmark\Executor\CompositeExecutor;
-use PhpBench\Benchmark\Executor\DebugExecutor;
+use PhpBench\Benchmark\Executor\Benchmark\DebugExecutor;
 use PhpBench\Benchmark\Executor\Method\RemoteMethodExecutor;
-use PhpBench\Benchmark\Executor\MicrotimeExecutor;
+use PhpBench\Benchmark\Executor\Benchmark\MicrotimeExecutor;
 use PhpBench\Benchmark\Metadata\AnnotationReader;
 use PhpBench\Benchmark\Metadata\Driver\AnnotationDriver;
 use PhpBench\Benchmark\Metadata\MetadataFactory;
