+++ tests/Unit/Benchmark/Executor/Benchmark/DebugExecutorTest.php
@@ -10,9 +10,9 @@
  *
  */
 
-namespace PhpBench\Tests\Unit\Benchmark\Executor;
+namespace PhpBench\Tests\Unit\Benchmark\Executor\Benchmark;
 
-use PhpBench\Benchmark\Executor\DebugExecutor;
+use PhpBench\Benchmark\Executor\Benchmark\DebugExecutor;
 use PhpBench\Benchmark\Metadata\SubjectMetadata;
 use PhpBench\Benchmark\Remote\Launcher;
 use PhpBench\Model\Iteration;
