+++ tests/Unit/Benchmark/Executor/Benchmark/benchmarks/static_method.tmp
@@ -1,0 +1,1 @@
+Static method method executed
