+++ tests/Unit/Benchmark/Executor/ParameterSetExtractorTest.php
@@ -56,8 +56,8 @@ class ParameterSetExtractorTest extends TestCase
     private function provideParams(array $providers)
     {
         $payload = new Payload(__DIR__ . '/../../../../lib/Benchmark/Remote/template/parameter_set_extractor.template', [
-            'bootstrap' => __DIR__ . '/benchmarks/ParamProviderBench.php',
-            'file' => __DIR__ . '/benchmarks/ParamProviderBench.php',
+            'bootstrap' => __DIR__ . '/Benchmark/benchmarks/ParamProviderBench.php',
+            'file' => __DIR__ . '/Benchmark/benchmarks/ParamProviderBench.php',
             'class' => ParamProviderBench::class,
             'paramProviders' => json_encode($providers),
         ]);
